JEAP-5487: Applied PR comments

+++ jeap-reaction-observer-domain/src/main/java/ch/admin/bit/jeap/reaction/observer/domain/ObservedReactionsAggregatedStatistics.java
@@ -1,4 +1,4 @@
 package ch.admin.bit.jeap.reaction.observer.domain;
 
-public record ObservedReactionsAggregatedStatistics(String component, String triggerType, String triggerFqn, String actionType, String actionFqn, int count, float median, float percentage) {
+public record ObservedReactionsAggregatedStatistics(String component, String triggerType, String triggerFqn, String actionType, String actionFqn, long count, double median, double percentage) {
 }
